requirements.txt
google_api_python_client==2.80.0
google_auth_oauthlib==0.7.1
pandas==1.3.3
protobuf==4.22.1
streamlit==1.12.2
